Contact components improvements
