+++ scripts/apps/contacts/components/ContactFooter.tsx
@@ -1,14 +1,18 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import {VersionCreated, SocialInfo} from 'apps/contacts/components/fields';
 import {getContactType} from '../../contacts/helpers';
 import classNames from 'classnames';
 import {gettext} from 'core/utils';
+import {IContact} from 'superdesk-api';
 
+interface IProps {
+    item: IContact;
+    svc: any;
+}
 /**
  * Contact footer - renders footer for contact card used in grid view
  */
-export const ContactFooter: React.StatelessComponent<any> = ({item, svc}) => {
+export const ContactFooter: React.FunctionComponent<IProps> = ({item, svc}) => {
     const cssClass = classNames(
         'sd-grid-item__footer sd-grid-item__footer--padded',
         {padded: getContactType(item) === 'organisation' || !item.organisation},
@@ -23,9 +27,4 @@ export const ContactFooter: React.StatelessComponent<any> = ({item, svc}) => {
             <SocialInfo item={item} />
         </div>
     );
-};
-
-ContactFooter.propTypes = {
-    item: PropTypes.object,
-    svc: PropTypes.object.isRequired,
 };
