+++ scripts/apps/contacts/components/ContactHeader.tsx
@@ -1,15 +1,12 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, {FunctionComponent} from 'react';
 import {gettext} from 'core/utils';
+import {IContact} from 'superdesk-api';
 
 /**
  * Contact header - renders header for contact card used in grid view
  */
-export const ContactHeader: React.StatelessComponent<any> = ({item}) => {
+export const ContactHeader: FunctionComponent<{item: IContact}> = ({item}) => {
     const typeTooltip = item.first_name ? gettext('Person Contact') : gettext('Organisation Contact');
-
-    const contactIcon = item.first_name ?
-        <i className="icon-user" /> : <i className="icon-business" />;
 
     return (
         <div
@@ -18,11 +15,7 @@ export const ContactHeader: React.StatelessComponent<any> = ({item}) => {
             data-sd-tooltip={item.public ? typeTooltip : typeTooltip + ' ' + gettext('(Private)')}
             data-flow="right"
         >
-            {contactIcon}
+            <i className={`${item.first_name ? 'icon-user' : 'icon-business'}`} />
         </div>
     );
-};
-
-ContactHeader.propTypes = {
-    item: PropTypes.object,
 };
