+++ scripts/apps/contacts/components/Form/ContactNumberInput.tsx
@@ -1,20 +1,34 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import {Row, LineInput, Input, SelectInput, Toggle, Label} from './';
+import {Row, LineInput, SelectInput, Input, Toggle} from './';
 import {KEYCODES} from '../../../contacts/constants';
 import {set, get, isEmpty} from 'lodash';
 import {gettext} from 'core/utils';
+import {Label} from 'superdesk-ui-framework/react';
 
-export class ContactNumberInput extends React.Component<any, any> {
-    static propTypes: any;
-    static defaultProps: any;
+interface IProps {
+    remove: () => void,
+    field: string,
+    value: Record<string, any>,
+    label: string,
+    onChange: (field: string, value: any) => void,
+    readOnly: boolean,
+    usages: Array<{ qcode: string }>,
+}
 
+interface IState {
+    preventSwitch: boolean;
+    touched: any;
+}
+
+export class ContactNumberInput extends React.Component<IProps, IState> {
     constructor(props) {
         super(props);
+
         this.state = {
             preventSwitch: false,
             touched: {},
         };
+
         this.onChange = this.onChange.bind(this);
         this.onBlur = this.onBlur.bind(this);
         this.isFieldInvalid = this.isFieldInvalid.bind(this);
@@ -40,6 +54,7 @@ export class ContactNumberInput extends React.Component<any, any> {
         } else {
             this.setState({preventSwitch: false});
         }
+
         this.props.onChange(field, value);
     }
 
@@ -59,12 +74,10 @@ export class ContactNumberInput extends React.Component<any, any> {
                 >
                     <Label text={gettext('Number')} />
                     <Input
-                        field={`${field}.number`}
                         value={get(value, 'number', '')}
                         onChange={onChange}
-                        onBlur={this.onBlur}
                         type="text"
-                        readOnly={readOnly}
+                        disabled={readOnly}
                     />
                 </LineInput>
                 <LineInput readOnly={readOnly} className="sd-line-input__usage">
@@ -90,40 +103,24 @@ export class ContactNumberInput extends React.Component<any, any> {
                     />
                 </LineInput>
                 <LineInput readOnly={readOnly}>
-                    {!readOnly &&
-                        (
-                            <a
-                                tabIndex={0}
-                                className="icn-btn sd-line-input__icon"
-                                onClick={remove}
-                                onKeyDown={(event) => {
-                                    if (event && event.keyCode === KEYCODES.ENTER) {
-                                        event.preventDefault();
-                                        remove();
-                                    }
-                                }}
-                            >
-                                <i className="icon-trash" />
-                            </a>
-                        )
-                    }
+                    {!readOnly && (
+                        <a
+                            tabIndex={0}
+                            className="icn-btn sd-line-input__icon"
+                            onClick={remove}
+                            onKeyDown={(event) => {
+                                if (event && event.keyCode === KEYCODES.ENTER) {
+                                    event.preventDefault();
+                                    remove();
+                                }
+                            }}
+                        >
+                            <i className="icon-trash" />
+                        </a>
+                    )}
                 </LineInput>
 
             </Row>
         );
     }
 }
-
-ContactNumberInput.propTypes = {
-    remove: PropTypes.func,
-    field: PropTypes.string,
-    value: PropTypes.object,
-    label: PropTypes.string,
-    onChange: PropTypes.func,
-    readOnly: PropTypes.bool,
-    usages: PropTypes.array,
-};
-
-ContactNumberInput.defaultProps = {
-    readOnly: false,
-};
